[Remove] test_command_read를 test_command에 Merge 후 제거
